AJAX requests now skips customError behavior
Added code i Global.asax to prevent redirects to ~/ or ~/404.cshtml
when the request is considered an AJAX request.
This is to prevent redirections of the requests which results in a
valid page and status code 200 and not the status code from thrown
HttpException which results in that the success/fail functionality
always thinks it's a success, even when it's not.

Fixes #104